refactor(remote): extract pinned worker SSH options
