+++ crates/horizon-core/src/lib.rs
@@ -23,6 +23,8 @@ pub mod remote_environment_observation;
 mod remote_hosts;
 pub mod remote_provider_config;
 pub mod remote_ssh_identity;
+#[cfg(target_os = "linux")]
+mod remote_worker_ssh;
 pub mod remote_worker_status;
 pub mod remote_workspace;
 pub mod remote_workspace_recovery;
